services/ReSys.ML/requirements.txt
# Server
fastapi
uvicorn[standard]
pydantic
python-multipart

# Machine Learning
torch
sentence-transformers
numpy

# OpenTelemetry

opentelemetry-api

opentelemetry-sdk

opentelemetry-instrumentation-fastapi

opentelemetry-exporter-otlp

opentelemetry-exporter-otlp-proto-grpc


